+++ packages/opentelemetry-sdk-trace-base/src/export/ReadableSpan.ts
@@ -31,6 +31,7 @@ export interface ReadableSpan {
   readonly kind: SpanKind;
   readonly spanContext: () => SpanContext;
   readonly parentSpanId?: string;
+  readonly parentSpanContext?: SpanContext;
   readonly startTime: HrTime;
   readonly endTime: HrTime;
   readonly status: SpanStatus;